third commit

+++ createMatch.php
@@ -7,8 +7,6 @@ $database = "sportodds";
 
 //create conncetion
 $conn = mysqli_connect($servername, $username, $password, $database);
-// echo "SELECT umpire_id, umpire_name FROM umpire WHERE umpire_id = $umpire_id";
-// exit;
 
 if (!$conn) {
     die("Sorry Connection unsuccessful" . mysqli_connect_error());
@@ -22,9 +20,22 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     $match_teamB = $_POST['match_teamB'];
     $match_type = $_POST['match_type'];
     $match_venue = $_POST['match_venue'];
+    $match_dt = $_POST['match_dt'];
+    $datetime = new DateTime($match_dt);
+    $dt = $datetime->format("Y-m-d H:i");
     $match_status = $_POST['match_status'];
 
-    $sql = "INSERT INTO `m_atch` (`umpire_id`, `match_teamA`, `match_teamB`, `match_type`, `match_venue`, `match_status`) VALUES ('$umpire_id', '$match_teamA', '$match_teamB', '$match_type', '$match_venue', '$match_status')";
+    $overs = 0;
+    if ($match_type == "T10") {
+        $overs = '10';
+    } elseif ($match_type == "T20") {
+        $overs = '20';
+    } elseif ($match_type == "ODI") {
+        $overs = '50';
+    }
+
+    $sql = "INSERT INTO `m_atch` (`umpire_id`, `match_teamA`, `match_teamB`, `match_type`, `overs`, `match_venue`, `match_status`, `match_dt`) 
+    VALUES ('$umpire_id', '$match_teamA', '$match_teamB', '$match_type', '$overs', '$match_venue', '$match_status', '$dt')";
     $result = mysqli_query($conn, $sql);
     if ($result) {
         echo '<div class="alert alert-success alert-dismissible fade show" role="alert">
@@ -38,6 +49,21 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
                 <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
                 </div>';
     }
+
+    $host = "127.0.0.1";
+    $port = "857";
+    $socket = socket_create(AF_INET, SOCK_STREAM, 0) or die("Not created");
+    socket_connect($socket, $host, $port) or die("Not connected");
+
+    $msg = $match_teamA;
+
+    socket_write($socket, $msg, strlen($msg));
+
+    $msg .= $match_teamB;
+
+    socket_write($socket, $msg, strlen($msg));
+
+    socket_close($socket);
 }
 
 ?>
@@ -63,13 +89,17 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     <link rel="stylesheet" href="https://cdn.datatables.net/1.10.24/css/jquery.dataTables.min.css">
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
     <script src="https://cdn.datatables.net/1.10.24/js/jquery.dataTables.min.js"></script>
+    <!-- Select 2 -->
+    <link href="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/css/select2.min.css" rel="stylesheet" />
+    <script src="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/js/select2.min.js"></script>
 
 </head>
 
 <body>
 
     <header>
-        <img src="assets/img/logo/logo.png" alt="logo">
+        <a href="admin.php"><img src="assets/img/logo/logo.png" alt="logo"></a>
+
         <nav>
             <ul>
                 <!-- <li><a href="#">Home</a></li>
@@ -81,8 +111,12 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
         </nav>
 
         <div class="btn">
-            <button class="button">Login</button>
-            <button class="button signup">Admin</button>
+            <a href="logout.php" class="cmn--btn" data-bs-toggle="" data-bs-target="">
+                <span>Logout</span>
+            </a>
+            <a href="#0" class="cmn--btn2" data-bs-toggle="" data-bs-target="">
+                <span>Admin</span>
+            </a>
         </div>
     </header>
     <!--Header End-->
@@ -102,6 +136,9 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
                 </li>
                 <li>
                     <a href="createUmpire.php">Add Umpire</a>
+                </li>
+                <li>
+                    <a href="createVenue.php">Add Venue</a>
                 </li>
                 <li>
                     <a class="active" href="createMatch.php">Add Match..</a>
@@ -128,7 +165,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
         </div>
 
         <div class="container fcontainer">
-            <table id="" class="table">
+            <table id="myTable" class="table">
                 <thead>
                     <tr>
                         <th scope="col">S No</th>
@@ -137,6 +174,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
                         <th scope="col">Team B</th>
                         <th scope="col">Match Type</th>
                         <th scope="col">Match Venue</th>
+                        <th scope="col">Match Date & Time</th>
                         <th scope="col">Match Status</th>
                         <th scope="col">Actions</th>
                     </tr>
@@ -144,36 +182,55 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
                 <tbody>
                     <?php
                     $srno = 0;
-                    $sql = "Select * from `m_atch`";
-                    $result = mysqli_query($conn, $sql);
+                    $query = "SELECT m.*, v.venue_name AS venue_name
+                    FROM m_atch AS m
+                    INNER JOIN venue AS v ON m.match_venue = v.venue_id
+                    ORDER BY m.match_id DESC";
+
+                    $result = mysqli_query($conn, $query);
+
                     if ($result) {
                         while ($row = mysqli_fetch_assoc($result)) {
                             $srno = $srno + 1;
                             $teamA = $row['match_teamA'];
                             $teamB = $row['match_teamB'];
+                            $umpir_id = $row['umpire_id'];
+                            $datetime = $row['match_dt'];
 
-                            $sql = "SELECT * FROM `team` where `team_id` = '$teamA'";
-                            $result1 = mysqli_query($conn, $sql);
-                            $row1 = mysqli_fetch_assoc($result1);
+                            $sqlUmpire = "SELECT * FROM `umpire` WHERE `umpire_id` = $umpir_id";
+                            $resultUmpire = mysqli_query($conn, $sqlUmpire);
+                            $rowUmpire = mysqli_fetch_assoc($resultUmpire);
 
+                            $sqlTeamA = "SELECT * FROM `team` WHERE `team_id` = '$teamA'";
+                            $resultTeamA = mysqli_query($conn, $sqlTeamA);
+                            $rowTeamA = mysqli_fetch_assoc($resultTeamA);
 
-                            $sql = "SELECT * FROM `team` where `team_id` = '$teamB'";
-                            $result2 = mysqli_query($conn, $sql);
-                            $row2 = mysqli_fetch_assoc($result2);
-                            //  print_r ($result2);exit;
+                            $sqlTeamB = "SELECT * FROM `team` WHERE `team_id` = '$teamB'";
+                            $resultTeamB = mysqli_query($conn, $sqlTeamB);
+                            $rowTeamB = mysqli_fetch_assoc($resultTeamB);
 
-                            echo "<tr>
-                            <td scope='row'>" . $srno . "</td>   
-                            <td>" . $row['umpire_id'] . "</td>
-                            <td>" . $row1['team_name'] . "</td>
-                            <td>" . $row2['team_name'] . "</td>
-                            <td>" . $row['match_type'] . "</td>
-                            <td>" . $row['match_venue'] . "</td>
-                            <td>" . $row['match_status'] . "</td>
-                            <td> 
-                            <button class='edit btn btn-sm btn-primary'><a href='updateMatch.php?updatematch_id=" . $row['match_id'] . "' class='text-light text-decoration-none'>Edit</a></button>
-                            <button class='delete btn btn-sm btn-danger'><a href='deleteMatch.php?deletematch_id=" . $row['match_id'] . "' class='text-light text-decoration-none' onclick='return deleteRecord(" . $row['match_id'] . ")'>Delete</a></button>
-                            </tr>";
+                            $dateTime = DateTime::createFromFormat('Y-m-d H:i:s', $datetime);
+
+                            if ($dateTime !== false) {
+                                $formattedDate = $dateTime->format('Y-m-d');
+                                $formattedTime = $dateTime->format('h:i:s A');
+
+                                echo "<tr>
+                                <td scope='row'>" . $srno . "</td>   
+                                <td>" . $rowUmpire['umpire_name'] . "</td>
+                                <td>" . $rowTeamA['team_name'] . "</td>
+                                <td>" . $rowTeamB['team_name'] . "</td>
+                                <td>" . $row['match_type'] . "</td>
+                                <td>" . $row['venue_name'] . "</td>
+                                <td>" . $formattedDate . " " . $formattedTime . "</td>
+                                <td>" . $row['match_status'] . "</td>
+                                <td> 
+                                    <button class='edit btn btn-sm btn-primary'><a href='updateMatch.php?updatematch_id=" . $row['match_id'] . "' class='text-light text-decoration-none'>Edit</a></button>
+                                    <button class='delete btn btn-sm btn-danger'><a href='deleteMatch.php?deletematch_id=" . $row['match_id'] . "' class='text-light text-decoration-none' onclick='return deleteRecord(" . $row['match_id'] . ")'>Delete</a></button>
+                                </tr>";
+                            } else {
+                                echo "Error: Unable to create DateTime object.";
+                            }
                         }
                     }
                     ?>
@@ -197,11 +254,9 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
         // let table = new DataTable('#myTable');
         $(document).ready(function() {
             $("#myTable").DataTable();
+            $('.select-2').select2();
         });
     </script>
-
-
-
 
 </body>
